requirements.txt
#
# ```
#   $ pip install --upgrade pip
# ```
#
# Then you probably want to work from your local PC:
#
# Start by cloning the project from github:
# ```
#   $ cd ~/Documents/
#   $ git clone https://github.com/Ayomide100/ai-in-medicine.git
#  
# ```
#
# or:
#
# You can download the `.zip` project files [here](https://github.com/Ayomide100/Ai-in-medicine)
# and extract the project files.
#
# ```
#   $ cd <your_download_dir>
#  
# ```
#
# Then install these requirements:
# ```
#   $ pip install --upgrade -r requirements.txt
# ```
#
# Finally, start the project:
# ```
#   $ python __init__.py
# ```
#

jinja2>=2.10
werkzeug>=0.14.1
flask>=0.12.2
numpy>=1.14.0
scipy>=1.0.0
pandas==0.22.0
scikit-learn==0.19.1
passlib==1.7.1
pillow==7.1.1
torch==1.4.0+cpu
torchvision==0.5.0+cpu
